backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
discord.py==2.3.2
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.4
pydantic==2.5.3
schedule==1.2.1
openai